Back to single (float) and fix rounding issue

+++ CppRuntime/CppRuntime.cpp
@@ -7,7 +7,7 @@
 
 extern "C" void __declspec(dllexport) __stdcall SomeCalculationsCPU
 (
-	double* a_h,
+	float* a_h,
 	const unsigned int N,
 	const unsigned int M
 )
@@ -15,7 +15,7 @@ extern "C" void __declspec(dllexport) __stdcall SomeCalculationsCPU
 	Timer sw;
 	printf("	Start unmanaged C %+" PRId64 " ms\n", sw.get_elapsed_time().count() / CLOCKS_PER_SEC / 1000);
 	for (unsigned int i = 0; i < N; i++)
-		*(a_h + i) = *(a_h + i) * *(a_h + i) * 0.1 - *(a_h + i) - 10;
+		*(a_h + i) = *(a_h + i) * *(a_h + i) * 0.1F - *(a_h + i) - 10.0F;
 	printf("	Unmanaged C stopped %+" PRId64 " ms\n", sw.get_elapsed_time().count() / CLOCKS_PER_SEC / 1000);
 }
 
@@ -29,7 +29,7 @@ void ResetCursor() {
 int main(void)
 {
 	clock_t start, stop;
-	double* a_h;
+	float* a_h;
 	const unsigned int N = 2048 * 2048 * 8 * 4;
 	const unsigned int M = 10;
 	const int cuBlockSize = 512;
@@ -39,9 +39,9 @@ int main(void)
 		Timer sw;
 		printf("Start C CPU\n");
 
-		size_t size = N * sizeof(double);
-		a_h = (double*)malloc(size);
-		for (unsigned i = 0; i < N; i++) *(a_h + i) = (double)i;
+		size_t size = N * sizeof(float);
+		a_h = (float*)malloc(size);
+		for (unsigned i = 0; i < N; i++) *(a_h + i) = (float)i;
 		printf(" Allocating %+" PRId64 " ms\n", sw.get_elapsed_time().count() / CLOCKS_PER_SEC / 1000);
 
 		printf(" Start computing %+" PRId64 " ms\n", sw.get_elapsed_time().count() / CLOCKS_PER_SEC / 1000);
